+++ pkg/ssh/resize_window_windows.go
@@ -1,0 +1,26 @@
+//go:build windows
+// +build windows
+
+package ssh
+
+func resizeWindow(session *ssh.Session) {
+	go func() {
+		prevWSize, err := dockerterm.GetWinsize(os.Stdout.Fd())
+		if err != nil {
+			oktetoLog.Infof("request for terminal size failed: %s", err)
+		}
+		for {
+			time.Sleep(time.Millisecond * 250)
+			ws, err := dockerterm.GetWinsize(os.Stdout.Fd())
+			if err != nil {
+				oktetoLog.Infof("request for terminal size failed: %s", err)
+			}
+			if prevWSize.Height != ws.Height || prevWSize.Width != ws.Width {
+				if err := dockerterm.SetWinsize(os.Stdout.Fd(), ws); err != nil {
+					oktetoLog.Infof("request for terminal resize failed: %s", err)
+				}
+			}
+			prevWSize = ws
+		}
+	}()
+}
